Refactor: ssl (#997)
* ssl_context is libevent-free so should be in net

* Tweak previous
