+++ src/libmeasurement_kit/net/ssl_context.hpp
@@ -1,19 +1,15 @@
 // Part of measurement-kit <https://measurement-kit.github.io/>.
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
-#ifndef SRC_LIBMEASUREMENT_KIT_LIBEVENT_SSL_CONTEXT_HPP
-#define SRC_LIBMEASUREMENT_KIT_LIBEVENT_SSL_CONTEXT_HPP
+#ifndef SRC_LIBMEASUREMENT_KIT_NET_SSL_CONTEXT_HPP
+#define SRC_LIBMEASUREMENT_KIT_NET_SSL_CONTEXT_HPP
 
 #include <measurement_kit/common.hpp>
 
 #include <openssl/ssl.h>
 
-// Forward declarations
-struct ssl_st;
-struct ssl_ctx_st;
-
 namespace mk {
-namespace libevent {
+namespace net {
 
 class SslContext : public NonCopyable, public NonMovable {
   public:
@@ -31,6 +27,6 @@ class SslContext : public NonCopyable, public NonMovable {
     ssl_ctx_st *ctx = nullptr;
 };
 
-} // namespace libevent
+} // namespace net
 } // namespace mk
 #endif
